Get details about an Ensembl identifier

This function retrieves information about one or more Ensembl identifiers. Ensembl identifiers for which information is available are: genes, exons, transcripts and proteins.

Usage

get_id(id, verbose = FALSE, warnings = TRUE, progress_bar = TRUE)

Arguments

id

A character vector of Ensembl identifiers. Ensembl identifiers have the form ENS[species prefix][feature type prefix][a unique eleven digit number]. id should not contain NAs. Please note that while 'ENSG00000157764' is a valid identifier as a query, 'ENSG00000157764.13' is not.

verbose

Whether to be verbose about the http requests and respective responses' status.

warnings

Whether to show warnings.

progress_bar

Whether to show a progress bar.

Value

A tibble of 9 variables:

id

Ensembl identifier.

id_latest

Ensembl identifier including the version suffix.

type

Entity type: gene ('Gene'), exon ('Exon'), transcript ('Transcript'), and protein ('Translation').

id_version

Ensembl identifier version, indicates how many times that entity has changed during its time in Ensembl.

release

Ensembl release version.

is_current

Is this the latest identifier for the represented entity.

genome_assembly_name

Code name of the genome assembly.

peptide

TODO

possible_replacement

TODO

Examples

get_id(c('ENSDARE00000830915', 'ENSG00000248378', 'ENSDART00000033574', 'ENSP00000000233'))
#> # A tibble: 0 × 9
#> # ℹ 9 variables: id <chr>, id_latest <chr>, type <chr>, id_version <int>,
#> #   release <int>, is_current <lgl>, genome_assembly_name <chr>, peptide <chr>,
#> #   possible_replacement <list>