Get the karyotype of a species

This function retrieves the set of chromosomes of a species.

Usage

get_karyotypes(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

species_name

The species name, i.e., the scientific name, all letters lowercase and space replaced by underscore. Examples: 'homo_sapiens' (human), 'ovis_aries' (Domestic sheep) or 'capra_hircus' (Goat).

verbose

Whether to be chatty.

warnings

Whether to print warnings.

progress_bar

Whether to show a progress bar.

Value

A tibble, each row being a chromosome, of 4 variables:

species_name

Ensembl species name: this is the name used internally by Ensembl to uniquely identify a species by name. It is the scientific name but formatted without capitalisation and spacing converted with an underscore, e.g., 'homo_sapiens'.

coord_system

Coordinate system type.

chromosome

Chromosome name.

length

Genomic length of the chromsome in base pairs.

Examples

# Get the karyotype of Caenorhabditis elegans
get_karyotypes('caenorhabditis_elegans')
#> # A tibble: 7 × 4
#>   species_name           coord_system chromosome   length
#>   <chr>                  <chr>        <chr>         <int>
#> 1 caenorhabditis_elegans chromosome   I          15072434
#> 2 caenorhabditis_elegans chromosome   II         15279421
#> 3 caenorhabditis_elegans chromosome   III        13783801
#> 4 caenorhabditis_elegans chromosome   IV         17493829
#> 5 caenorhabditis_elegans chromosome   V          20924180
#> 6 caenorhabditis_elegans chromosome   X          17718942
#> 7 caenorhabditis_elegans chromosome   MtDNA         13794

# Get the karyotype of the Giant panda
get_karyotypes('ailuropoda_melanoleuca')
#> # A tibble: 21 × 4
#>    species_name           coord_system     chromosome    length
#>    <chr>                  <chr>            <chr>          <int>
#>  1 ailuropoda_melanoleuca primary_assembly 1          212770937
#>  2 ailuropoda_melanoleuca primary_assembly 2          199809881
#>  3 ailuropoda_melanoleuca primary_assembly 3          147627920
#>  4 ailuropoda_melanoleuca primary_assembly 4          144794249
#>  5 ailuropoda_melanoleuca primary_assembly 5          130990914
#>  6 ailuropoda_melanoleuca primary_assembly 6          131587389
#>  7 ailuropoda_melanoleuca primary_assembly 7          141533052
#>  8 ailuropoda_melanoleuca primary_assembly 8          129245720
#>  9 ailuropoda_melanoleuca primary_assembly 9          103688518
#> 10 ailuropoda_melanoleuca primary_assembly 10         110580333
#> # ℹ 11 more rows