Retrieve variant sources

This function retrieves variant sources, i.e. a list of databases used by Ensembl from which variant information is retrieved.

Usage

get_variation_sources(
  species_name = "human",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)

Arguments

species_name

The species name, i.e., the scientific name, all letters lowercase and space replaced by underscore. Examples: 'homo_sapiens' (human), 'ovis_aries' (Domestic sheep) or 'capra_hircus' (Goat).

verbose

Whether to be chatty.

warnings

Whether to print warnings.

progress_bar

Whether to show a progress bar.

Value

A tibble, each row being a variant database, of 8 variables:

species_name

Ensembl species name: this is the name used internally by Ensembl to uniquely identify a species by name. It is the scientific name but formatted without capitalisation and spacing converted with an underscore, e.g., 'homo_sapiens'.

db_name

Database name.

type

Database type, e.g., chip (genotyping chip) or lsdb (locus-specific database).

version

Database version.

somatic_status

Somatic status.

description

Database description.

url

Database's URL.

data_types

Data types to be found at database.

Ensembl REST API endpoints

get_variation_sources makes GET requests to info/variation/:species.

Examples

# Retrieve variant sources for human (default)
get_variation_sources()
#> # A tibble: 37 × 8
#>    species_name db_name           type  version somatic_status description url  
#>    <chr>        <chr>             <chr> <chr>   <chr>          <chr>       <chr>
#>  1 human        dbSNP             NA    154     mixed          Variants (… http…
#>  2 human        Archive dbSNP     NA    154     mixed          Former dbS… http…
#>  3 human        dbSNP HGVS        NA    154     mixed          HGVS annot… http…
#>  4 human        Former dbSNP      NA    154     mixed          Former dbS… http…
#>  5 human        PharmGKB          NA    28/02/… germline       A pharmaco… http…
#>  6 human        HGMD-PUBLIC       NA    2020.4  germline       Variants f… http…
#>  7 human        DGVa              NA    01/2020 mixed          Database o… http…
#>  8 human        NHGRI-EBI GWAS c… NA    03/03/… germline       Variants a… http…
#>  9 human        EGA               NA    03/2023 germline       Variants i… http…
#> 10 human        UniProt           NA    28/02/… mixed          Variants w… http…
#> # ℹ 27 more rows
#> # ℹ 1 more variable: data_types <list>

# Retrieve variant sources for mouse
get_variation_sources(species_name = 'mus_musculus')
#> # A tibble: 6 × 8
#>   species_name db_name type  version somatic_status description url   data_types
#>   <chr>        <chr>   <chr> <chr>   <chr>          <chr>       <chr> <list>    
#> 1 mus_musculus dbSNP   NA    150     mixed          Variants (… http… <chr [1]> 
#> 2 mus_musculus Archiv… NA    150     mixed          Former var… http… <chr [1]> 
#> 3 mus_musculus DGVa    NA    05/2018 mixed          Database o… http… <chr [2]> 
#> 4 mus_musculus IMPC    NA    28/11/… germline       Internatio… http… <chr [1]> 
#> 5 mus_musculus MGI     NA    28/11/… germline       Mouse Geno… http… <chr [1]> 
#> 6 mus_musculus EVA     NA    4       germline       Short vari… http… <chr [1]>